Add ListAll, ListProperty, ListAppointment tests
